Some small updates, changed image display style. Good results with imshow. Now can set vmin/vmax to a range to control which portions of the image get through.
I need to review the method used to color the images, I believe 0-100 is a scale of 0-100% along the a representation of the distribution. IE, 90-100 May represent the top 10% of the data. I believe the coloring may be done via interpolation, otherwise it's statistical analysis.

Right now vmin is 80 just for demo