Updates to sero and serodraw. Included methods for plotting counters, updated histogram plotting method. Refactored serodraw names to be capitalized
Fixed pixel ops in sero.py, issue was that had forgot to update col,row when advancing to the next pixel. Now does a good job and grouping in a single pass, image included in previous commit which contains results so far.

Now need to do second time over, to meld together id_groups that are touching. The is a good example of this in the image, where one of the centeral, larger blobs is split into two groups of ids. Presumably, this is because 2 groups of pixels got started from within the blob, and were both expanding before one could dominate the blob.

This should be addressable in second iteration, and could also be addressed by having a rigid order to check the pixels or making recursive calls when applying blob_ids.