Major work and bug fixes on the hungarian method for determining the best way to match edge_pixels between two blob2d's which are potentially part of the same blob3d.
Need to do a few more tests, especially as sources of information are somewhat flakey. Currently appears to be working, but will confirm with next commit. Note that there are two implentations for adjusting a matrix with all zeros covered by less than n-lines; there's a 'wikihow' and 'harvard' approach, each of which is currently linked to in code, and should be equivalent by a theorem, however are as so to compare results step by step against the limited examples that are offered.

Have not implemented the final step, where once fully derived cost_array is ready, we pick n-dim (from nxn matrix) zeros, such that there is one per row and column.