Updated to include visualization for the stitching between the optimally connected points of 2d blobs.
Note that this is a un-clean copy, and so currently contains a lot of debug output.

Next up:
Get a system for pickling setup; fails under current system or plots slides excessively close for unknown reasons.

After which, can give this a try on the real thing. Pickling is important, as even if the Hungarian Algorithm is optimal, it is still O(n^3)

Code cleanup planned; moving old methods to a separate file, and cleaning up formatting, naming conventions, etc etc.