Introduced Stitches, which is a class that serves to hold relevant subsets of the edge_pixels from blobs. Have updated plotBlobsVC to be able to plot these stitches and their respective lines.
Had been having issues with each slides id_num, which should have been remedied.
However, there is still some sort of id issue, as currently there is an issue with getIdLists (which uses ids), regardless of remapping. It seems to work under the test cases, but failed when presented with the actual scans.

Again, THIS COMMIT HAS AN ISSUE THAT WAS NOT PRESENT IN THE PREVIOUS COMMIT, RELATED TO THE BLOB IDS ASSIGNED TO PIXELS